requirements.txt
Django==3.1.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0
django-jazzmin==2.4.1
django-extensions==3.1.0
pyyaml==5.3.1
dataclasses-json==0.5.2
django-debug-toolbar==3.2
djangorestframework==3.12.2
drf-flex-fields==0.8.8
djangorestframework-gis==0.16
django-filter==2.4.0
sentry-sdk==0.19.5
django-leaflet==0.27.1
